pyproject.toml
[project]
name = "qiitaarticles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "pyperclip>=1.9.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0.0",
    "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
